refactored stakeOnValidator.
